Added pin connections for hall effect sensor and re-organized barometer and hall effect sensor esp32 folders
