+++ sensor_testing/esp32_files/hall_effect_sensor/platformio.ini
@@ -12,3 +12,5 @@
 platform = espressif32
 board = esp32dev
 framework = arduino
+monitor_speed = 115200
+; monitor_filters = esp32_exception_decoder
