+++ sensor_testing/schematics/drawings/2T3-2T4_Schematic.png
