+++ sensor_testing/schematics/drawings/Hall_Effect_Sensor_Schematic.jpg
