+++ sensor_testing/schematics/pin_connections/esp32_hall_effect_sensor.txt
@@ -1,0 +1,8 @@
+Hall Effect Sensor Pin Connections on the ESP32:
+
+- Hall Effect Sensor Left Pin (+) --> 5V
+- Hall Effect Sensor Middle Pin (-) --> GND
+- Hall Effect Sensor Right Pin --> GPIO23
+
+- LED Cathode (-) --> GND
+- LED Anode (+) --> GPIO2
